% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_plot_cvR2.R
\name{ab_plot_cvR2}
\alias{ab_plot_cvR2}
\title{Plots of cross-validated R-squared}
\usage{
ab_plot_cvR2(cvsl, X, col = "blue",
  ylab = "V-fold Cross Validated R-squared", title = NULL, ylim = c(0, 1))
}
\arguments{
\item{cvsl}{An object of class \code{CV.SuperLearner}}

\item{X}{The data.frame of features used in the CV.SuperLearner prediction}

\item{col}{(optional) Colors for plotting}

\item{ylab}{Name of the loss function (defaults uses generic "Risk") but it can be modified to be more specific, such as MSE, AUC, etc...}

\item{title}{Title for the plot (optional)}

\item{ylim}{set Y-axis limits, defaults to c(0,1)}
}
\description{
#' Plotting routine for cross-validated R-squared of super learner predictions of antibody data (along with its constituent algorithms)
}

