% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_tmle.R
\name{ab_tmle}
\alias{ab_tmle}
\title{Targeted maximum liklihood estimation for antibody measurements}
\usage{
ab_tmle(Y, Age, X = NULL, W = NULL, id = NULL, family = "gaussian",
  SL.library = c("SL.mean", "SL.glm", "SL.gam", "SL.loess"), diff = FALSE,
  RFnodesize = NULL, gamdf = NULL)
}
\arguments{
\item{Y}{antibody measurement}

\item{Age}{individual's age}

\item{X}{comparison group  (must be binary, 0/1, for tmle() ).}

\item{W}{matrix of covariates (optional)}

\item{id}{An optional cluster or repeated measures id variable. For cross-validation splits, \code{id} forces observations in the same cluster or for the same individual to be in the same validation fold.}

\item{family}{Model family (gaussian for continuous outcomes, binomial for binary outcomes)}

\item{SL.library}{Library of algorithms to include in the ensemble (see the \code{\link[SuperLearner]{SuperLearner}} package for details).}

\item{RFnodesize}{Optional argument to specify a range of minimum node sizes for the random Forest algorithm. If \code{SL.library} includes \code{SL.randomForest}, then the default is to search over node sizes of 15,20,...40. Specifying this option will override the default.}

\item{gamdf}{Optional argument to specify a range of degrees of freedom for natural smoothing splines in a generalized additive model. If \code{SL.library} includes \code{SL.gam}, then the default is to search over a range of df=2-6. Specifying this option will override the default.}
}
\description{
Targeted maximum liklihood estimation (TMLE) for mean antibody levels or differences in antibody levels
}
\details{
TBD
}
\examples{
TBD
}

