% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{design_matrix}
\alias{design_matrix}
\title{Convert a vector, matrix or data frame with factors into a design matrix}
\usage{
design_matrix(W)
}
\arguments{
\item{W}{A vector, matrix, or data.frame that includes numeric or factor variables.}
}
\value{
A design matrix version of \code{W} where factor variables have been converted into columns of indicator variables with the first level excluded.
}
\description{
An internal, convenience function that automatically transforms a vector, matrix, or data.frame with factors into a design matrix with indicator variables and an ommitted category.
}
\examples{

}

