% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garki_sero.R
\docType{data}
\name{garki_sero}
\alias{garki_sero}
\title{DESCRIPTION OF DATA HERE (what it is)}
\format{dataframe}
\source{
\href{https://osf.io/7frw2/}{Open Science Framework Storage}


          #----------TODO: Put examples here ---------------
}
\usage{
data(garki_sero)
}
\description{
DESCRIPTION OF DATA HERE (where its from, what kinds of things at the high-level are contained in it)
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{
iplotCurves(phe, times, phe[,c(61,121)], phe[,c(121,181)],
            chartOpts=list(curves_xlab="Time (hours)", curves_ylab="Root tip angle (degrees)",
                           scat1_xlab="Angle at 2 hrs", scat1_ylab="Angle at 4 hrs",
                           scat2_xlab="Angle at 4 hrs", scat2_ylab="Angle at 6 hrs"))}
}
\references{
Molineaux L, Gramiccia G, 1980. The Garki Project. Geneva: World Health Organisation.
http://whqlibdoc.who.int/publications/9241560614.pdf
}
\keyword{#----------TODO:}
\keyword{Edit}
\keyword{I}
\keyword{OSF}
\keyword{README}
\keyword{datasets}
\keyword{found}
\keyword{here.}
\keyword{if}
\keyword{in}
\keyword{like---------------}
\keyword{on}
\keyword{put}
\keyword{references}
\keyword{the}
\keyword{you'd}

