% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti_enterics.R
\docType{data}
\name{haiti_enterics}
\alias{haiti_enterics}
\title{Open Science Framework data on enteric disease in Haiti}
\format{dataframe}
\source{
\href{https://osf.io/7frw2/}{Open Science Framework Storage}


          #----------TODO: Put examples here ---------------
}
\usage{
data(haiti_enterics)
}
\description{
The Haiti data include measurements from 142 children measured between 1 and 9 times in Leogane, Haiti.
The codebooks for each file include variable descriptions and details.
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{
iplotCurves(phe, times, phe[,c(61,121)], phe[,c(121,181)],
            chartOpts=list(curves_xlab="Time (hours)", curves_ylab="Root tip angle (degrees)",
                           scat1_xlab="Angle at 2 hrs", scat1_ylab="Angle at 4 hrs",
                           scat2_xlab="Angle at 4 hrs", scat2_ylab="Angle at 6 hrs"))}
}
\references{
Moore CR, Johnson LS, Kwak I-Y, Livny M, Broman KW,
Moss, Delynn M., Jeffrey W. Priest, Kathy Hamlin, Gordana Derado, Joel Herbein, William A. Petri Jr, and Patrick J. Lammie. 2014.
“Longitudinal Evaluation of Enteric Protozoa in Haitian Children by Stool Exam and Multiplex Serologic Assay.”
The American Journal of Tropical Medicine and Hygiene 90 (4): 653–60.
}
\keyword{#----------TODO:}
\keyword{Edit}
\keyword{I}
\keyword{OSF}
\keyword{README}
\keyword{datasets}
\keyword{found}
\keyword{here.}
\keyword{if}
\keyword{in}
\keyword{like---------------}
\keyword{on}
\keyword{put}
\keyword{references}
\keyword{the}
\keyword{you'd}

