% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miton_malaria.R
\docType{data}
\name{miton_malaria}
\alias{miton_malaria}
\title{dataset: Miton, Haiti malaria (MSP-1) serology data}
\format{A data frame with 383 observations and 3 variables.
\describe{
 \item{\code{id}}{individual ID}
 \item{\code{age}}{age in years. Note that age records years completed except for children <1.}
 \item{\code{msp13d7}}{reponse (MFI-bg) to the Plasmodium falciparum MSP-1_19 fusion protein cloned from P. falciparum isolate 3D7}
 }}
\usage{
data(miton_malaria)
}
\description{
A 1998 cross-sectional serological survey in the community of Miton, Haiti, near Leogane, collected finger prick blood specimens from 383 people ages 2 weeks to 90 years old as part of a trial to measure the effect of salt fortified with diethylcarbamazine on lymphatic filariasis transmission (Freeman et al. 2001). Antibody levels were measured on the Luminex platform and are in units of median flourescence intensity minus background (MFI-bg).
}
\references{
B. F. Arnold et al., Serological measures of malaria transmission in haiti: comparison of longitudinal and cross-sectional methods. PLoS One. 9, e93684 (2014).

A. R. Freeman et al., A community-based trial for the control of lymphatic filariasis and iodine deficiency using salt fortified with diethylcarbamazine and iodine. Am. J. Trop. Med. Hyg. 65, 865–871 (2001).
}
\keyword{datasets}
