% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.Yman2016.R
\name{predict.SL.Yman2016}
\alias{predict.SL.Yman2016}
\title{Internal prediction method for SL.Yman2016}
\usage{
\method{predict}{SL.Yman2016}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{SL.Yman2016}}

\item{newdata}{New data used for predicting the outcome (must be univariate and equal to age, or if multivariate, it must include a variable 'Age')}

\item{...}{Other arguments passed to the function (currently ignored)}
}
\value{
pred Predicted values given \code{newdata}
}
\description{
Internal prediction method for SL.Yman2016
}

