% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_cvRF.R
\name{SLAb.cvRF}
\alias{SLAb.cvRF}
\title{Selects optimal tree depth for randomForest() using corss-validation by tuning the nodesize parameter}
\usage{
SLAb.cvRF(Y, X, id, SLlib, print = FALSE)
}
\arguments{
\item{Y}{outcome}

\item{X}{matrix of features that predict Y}

\item{id}{id variable to identify independent observations for V-splits}

\item{SLlib}{SuperLearner library}

\item{print}{logical. print messages? Defaults to FALSE}
}
\value{
returns a list with updated SuperLearner library, the optimal node size, and cvRisks
}
\description{
Selects optimal tree depth for randomForest() using corss-validation by tuning the nodesize parameter
}
\examples{

}

