% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_cvGAM.R
\name{slab_cvGAM}
\alias{slab_cvGAM}
\title{Selects optimal degrees of freedom for SL.gam() using cross-validation by tuning the df parameter}
\usage{
slab_cvGAM(Y, X, id = NULL, SL.library, print = FALSE, df = 2:10)
}
\arguments{
\item{Y}{The outcome. Must be a numeric vector.}

\item{X}{A matrix of features that predict Y, usually a data.frame.}

\item{id}{An optional cluster or repeated measures id variable. For cross-validation splits, \code{id} forces observations in the same cluster or for the same individual to be in the same validation fold.}

\item{SL.library}{SuperLearner library}

\item{print}{logical. print messages? Defaults to FALSE}

\item{df}{a sequence of degrees of freedom to control the smoothness of natural splines in the GAM model. Defaults to a sequence from 2 to 10}
}
\value{
returns a list with updated SuperLearner library, the optimal node size, and cvRisks
}
\description{
Selects optimal degrees of freedom for SL.gam() using cross-validation by tuning the df parameter
}
\details{
\code{slab_cvGAM} is an internal function called by \code{\link[SLAb]{slab_curve}} or \code{slab_tmle} if SL.gam() is included in the algorithm library. It performs an addition pre-screen step of selecting the optimal node depth for random forest using cross validation. The default range of degrees of freedom is 2, 3, ...10, which is usually sufficient.  In the context of Age-antibody curves, without this tuning step the default parameters for GAM will usually under-smooth. This additional selection step tunes the smoothing parameter. Cross-validated risks are estimated using \code{\link[SuperLearner]{SuperLearner}}.
}
\examples{
TBD
}

