% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_cvRF.R
\name{slab_cvRF}
\alias{slab_cvRF}
\title{Selects optimal tree depth for SL.randomForest() using cross-validation by tuning the nodesize parameter}
\usage{
slab_cvRF(Y, X, id, SL.library, print = FALSE)
}
\arguments{
\item{Y}{The outcome. Must be a numeric vector.}

\item{X}{A matrix of features that predict Y, usually a data.frame.}

\item{id}{An optional cluster or repeated measures id variable. For cross-validation splits, \code{id} forces observations in the same cluster or for the same individual to be in the same validation fold.}

\item{SL.library}{SuperLearner library}

\item{print}{logical. print messages? Defaults to FALSE}
}
\value{
returns a list with updated SuperLearner library, the optimal node size, and cvRisks
}
\description{
Selects optimal tree depth for SL.randomForest() using cross-validation by tuning the nodesize parameter
}
\details{
\code{slab_cvRF} is an internal function called by \code{\link[SLAb]{slab_curve}} or \code{slab_tmle} if SL.randomForest() is included in the algorithm library. It performs an addition pre-screen step of selecting the optimal node depth for random forest using cross validation. The default range of node sizes evaluated is 15, 20, ..., 40.  In the context of Age-antibody curves, without this tuning step random forest will fit extremely jagged curves that are clear overfits. This additional selection step prevents overfitting. Cross-validated risks are estimated using \code{\link[SuperLearner]{SuperLearner}}.
}
\examples{
TBD
}

