% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_cvSL.R
\name{slab_cvSL}
\alias{slab_cvSL}
\title{SLAb wrapper function to estimate CV risk for the super learner ensemble predicting antibody measurements using age and potentially other covariates}
\usage{
slab_cvSL(Y, Age, W = NULL, id = 1:length(Y), family = gaussian(),
  V = 10, SL.library = c("SL.mean", "SL.glm", "SL.bayesglm", "SL.loess",
  "SL.gam", "SL.randomForest"))
}
\arguments{
\item{Y}{log antibody measurement}

\item{SL.library}{}
}
\description{
SLAb wrapper function to estimate CV risk for the super learner ensemble predicting antibody measurements using age and potentially other covariates
}

