% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_plot_cvSL.R
\name{slab_plot_cvSL}
\alias{slab_plot_cvSL}
\title{Plotting routine for cross-validated risk of super learner predictions of antibody data (along with its constituent algorithms)}
\usage{
slab_plot_cvSL(x, col = "blue", title = NULL, xlim = NULL)
}
\arguments{
\item{x}{Object of class CV.SuperLearner}

\item{col}{(optional) Colors for plotting}

\item{title}{(optional) Title for the plot}

\item{xlim}{(optional) X-axis limits}
}
\description{
Plotting routine for cross-validated risk of super learner predictions of antibody data (along with its constituent algorithms)
}

