% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_tmle.R
\name{slab_tmle}
\alias{slab_tmle}
\title{slab_tmle}
\usage{
slab_tmle(Y, Age, X = NULL, W = NULL, id, SLlib = c("SL.mean", "SL.glm",
  "SL.loess", "SL.gam", "SL.randomForest"), diff = FALSE)
}
\arguments{
\item{Y}{antibody measurement}

\item{Age}{individual's age}

\item{X}{comparison group  (must be binary, 0/1, for tmle() ).}

\item{W}{matrix of covariates (optional)}

\item{id}{ID variable, unique for each individual (to identify repeated obs)}

\item{diff}{logical. calculate the difference between treatment groups? If FALSE (default) it returns the mean}

\item{SL.library}{SuperLearner algorithm library}
}
\description{
Targeted maximum liklihood estimation (TMLE) for mean antibody levels or differences in antibody levels
}
\details{
TBD
}
\examples{
TBD
}

