% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleAb.R
\name{tmleAb}
\alias{tmleAb}
\title{Targeted maximum liklihood estimation for antibody measurements}
\usage{
tmleAb(Y, X = NULL, W = NULL, id = NULL, SL.library = c("SL.mean",
  "SL.glm", "SL.gam", "SL.loess"), g.SL.library = c("SL.glm"), V = 5,
  RFnodesize = NULL, gamdf = NULL)
}
\arguments{
\item{Y}{antibody measurement}

\item{X}{comparison group  (must be binary, 0/1). If \code{X=NULL}, then the function returns the mean (rather than the difference between levels of \code{X}).}

\item{W}{matrix of covariates -- should probably at minimum include the individual's age (if available).}

\item{id}{An optional cluster or repeated measures id variable. For cross-validation splits, \code{id} forces observations in the same cluster or for the same individual to be in the same validation fold.}

\item{SL.library}{Library of models/algorithms to include in the ensemble for the outcome (see the \code{\link[SuperLearner]{SuperLearner}} package for details).}

\item{g.SL.library}{Optional library of models/algorithms to model group assignment. Default is to use main terms logistic regression (SL.glm).}

\item{V}{Number of cross-validation folds for Super Learning to estimate outcome (Q) and treatment (g) models (default is \code{V=5}).}

\item{RFnodesize}{Optional argument to specify a range of minimum node sizes for the random Forest algorithm. If \code{SL.library} includes \code{SL.randomForest}, then the default is to search over node sizes of 15,20,...40. Specifying this option will override the default.}

\item{gamdf}{Optional argument to specify a range of degrees of freedom for natural smoothing splines in a generalized additive model. If \code{SL.library} includes \code{SL.gam}, then the default is to search over a range of df=2-10. Specifying this option will override the default.}
}
\value{
\code{psi} Mean (if \code{X=NULL}) or difference

\code{se} Standard error of \code{psi}, estimated from the influence curve

\code{lb} Lower bound of the 95 percent confidence interval of \code{psi}

\code{ub} Upper bound of the 95 percent confidence interval of \code{psi}

\code{p} P-value for a test that \code{psi=0}

\code{tmle_fit} The original \code{tmle()} fit (see the \code{\link[tmle]{tmle}} package for details).
}
\description{
Targeted maximum liklihood estimation (TMLE) for mean antibody measurements or difference in antibody measurements between groups
}
\details{
The \code{tmleAb} function estimates adjusted means or differences in means in antibody measurements using targeted maximum likelihood estimation (TMLE). \code{tmleAb} assumes a continuous outcome (\code{family='gaussian'}), but if a binary outcome is passed to the function it will estimate means or differences in terms of seroprevalence.
}
\examples{
\dontrun{
# load the Garki project serology data, subset to round 5
data("garki_sero")
garki_sero$village <- factor(garki_sero$village)
garki_sero$sex <- factor(garki_sero$sex)
garki_sero$tr01 <- ifelse(garki_sero$tr=="Intervention",1,0)
d <- subset(garki_sero,serosvy==5)

# control and intervention village measurements
dc <- subset(d, tr=="Control")
di <- subset(d,tr=="Intervention")

# estimate means in control and intervention villages
# set a seed for perfectly reproducible splits
# in the V-fold cross validation
set.seed(12345)
cmean <-tmleAb(Y=log10(dc$ifatpftitre+1),
               W=dc[,c("ageyrs","sex","village")],
               id=dc$id)
set.seed(12345)
imean <-tmleAb(Y=log10(di$ifatpftitre+1),
               W=di[,c("ageyrs","sex","village")],
               id=di$id)

# estimate targeted difference in means
# bewteen control and intervention villages
# adjusted for age, sex and village
set.seed(12345)
psi_diff <-tmleAb(Y=log10(d$ifatpftitre+1),
                  X=d$tr01,
                  W=d[,c("ageyrs","sex","village")],
                  id=d$id)
}

}
\references{
Gruber S, van der Laan M. tmle: An R Package for Targeted Maximum Likelihood Estimation. J Stat Softw. 2012;51: 1–35.

van der Laan MJ, Polley EC, Hubbard AE. Super Learner. Stat Appl Genet Mol Biol. 2007;6: 1544–6115.
}
\seealso{
\code{\link[tmleAb]{agecurveAb}}, \code{\link[SuperLearner]{SuperLearner}},  \code{\link[tmle]{tmle}}
}
