% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miton_malaria.R
\docType{data}
\name{miton_malaria}
\alias{miton_malaria}
\title{DESCRIPTION OF DATA HERE (what it is)}
\format{dataframe
each row includes variable descriptions and details for a single individual}
\source{
\href{https://osf.io/7frw2/}{Open Science Framework Storage}


          #----------TODO: Put examples here ---------------
}
\usage{
data(miton_malaria)
}
\description{
DESCRIPTION OF DATA HERE (where its from, what kinds of things at the high-level are contained in it)
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{
iplotCurves(phe, times, phe[,c(61,121)], phe[,c(121,181)],
            chartOpts=list(curves_xlab="Time (hours)", curves_ylab="Root tip angle (degrees)",
                           scat1_xlab="Angle at 2 hrs", scat1_ylab="Angle at 4 hrs",
                           scat2_xlab="Angle at 4 hrs", scat2_ylab="Angle at 6 hrs"))}
}
\references{
Arnold, Benjamin F., Jeffrey W. Priest, Katy L. Hamlin, Delynn M. Moss, John M. Colford Jr, and Patrick J. Lammie. 2014.
“Serological Measures of Malaria Transmission in Haiti: Comparison of Longitudinal and Cross-Sectional Methods.” PloS One 9 (4): e93684.
}
\keyword{#----------TODO:}
\keyword{Edit}
\keyword{I}
\keyword{OSF}
\keyword{README}
\keyword{datasets}
\keyword{found}
\keyword{here.}
\keyword{if}
\keyword{in}
\keyword{like---------------}
\keyword{on}
\keyword{put}
\keyword{references}
\keyword{the}
\keyword{you'd}

